+++ grand_trade_auto/model/__init__.py
@@ -1,5 +1,6 @@
 # pylint: disable=missing-module-docstring
 __all__ = [
         'datafeed_src',
+        'exchange',
         'model_meta',
 ]