+++ tests/unit/orm/test_postgres_orm.py
@@ -120,6 +120,7 @@ def _test_create_schema(orm, test_func, table_name, table_schema='public',
 # Order of parameters must match order in _create_schemas() due to dependencies
 @pytest.mark.parametrize('method_name, table_name', [
     ('_create_schema_datafeed_src', 'datafeed_src'),
+    ('_create_schema_exchange', 'exchange'),
 ])
 def test__create_schemas_individually(pg_test_orm, method_name, table_name):
     """